MusicDemo

+++ MusicDemo/src/Servlet/DeleteSelMusicServlet.java
@@ -3,6 +3,7 @@ package Servlet;
 import DAO.MusicDao;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import entity.Music;
+import entity.User;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -14,6 +15,8 @@ import java.io.IOException;
 import java.util.HashMap;
 import java.util.Map;
 
+
+//批量删除
 @WebServlet("/deleteSelMusicServlet")
 
 public class DeleteSelMusicServlet extends HttpServlet {
@@ -25,40 +28,48 @@ public class DeleteSelMusicServlet extends HttpServlet {
 
         String[] values = req.getParameterValues("id[]");
 
+        User user = (User)req.getSession().getAttribute("user");
+        String userName = user.getUsername();
+
         MusicDao musicDao = new MusicDao();
         int count =0;
         Map<String,Object> return_map = new HashMap<>();
 
-        //数组中是所有要删除的id
-        for (int i = 0; i <values.length ; i++) {
-            int id = Integer.parseInt(values[i]);
-            Music music = musicDao.findMusicById(id);
+        if(userName.equals("admin")){
+            //数组中是所有要删除的id
+            for (int i = 0; i <values.length ; i++) {
+                int id = Integer.parseInt(values[i]);
+                Music music = musicDao.findMusicById(id);
 
-            int ret = musicDao.deleteMusicById(id);
-            if(ret == 1){
-                //删除服务器上的
-                File file = new File("D:\\javaProgram\\JAVA\\MusicDemo\\web\\"+music.getUrl()+".mp3");
-                if(file.delete()){
-                    count+=ret;
+                int ret = musicDao.deleteMusicById(id);
+                if(ret == 1){
+                    //删除服务器上的
+                    File file = new File("D:\\javaProgram\\JAVA\\MusicDemo\\web\\"+music.getUrl()+".mp3");
+                    if(file.delete()){
+                        count+=ret;
+                    }else {
+                        return_map.put("msg",false);
+                    }
                 }else {
                     return_map.put("msg",false);
+
                 }
+            }
+            if(count == values.length){
+                return_map.put("msg",true);
+
             }else {
                 return_map.put("msg",false);
 
             }
-        }
-        if(count == values.length){
-            return_map.put("msg",true);
 
         }else {
             return_map.put("msg",false);
-
         }
-
         //将map返回给前端
         ObjectMapper mapper = new ObjectMapper();
         mapper.writeValue(resp.getWriter(),return_map);
+
 
     }
 }
