+++ MusicDemo/src/Servlet/DeleteServlet.java
@@ -3,6 +3,7 @@ package Servlet;
 import DAO.MusicDao;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import entity.Music;
+import entity.User;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
@@ -21,33 +22,45 @@ public class DeleteServlet extends HttpServlet {
         req.setCharacterEncoding("utf-8");
         resp.setContentType("application/json;charset=utf-8");
 
+        User user = (User)req.getSession().getAttribute("user");
+        String userName = user.getUsername();
+
         String idStr = req.getParameter("id");
         int id = Integer.parseInt(idStr);
         MusicDao musicDao = new MusicDao();
 
-        Music music = musicDao.findMusicById(id);
-        if(music == null){
-            return;
-
-        }
         Map<String,Object> return_map = new HashMap<>();
         //msg    true
-        int ret = musicDao.deleteMusicById(id);
+        if(userName.equals("admin")){
+            Music music = musicDao.findMusicById(id);
+            if(music == null){
+                return;
+
+            }
+            int ret = musicDao.deleteMusicById(id);
             //数据库删除还要删除服务器的
-        if(ret == 1){
+            if(ret == 1){
                 //删除服务器上的
-            File file = new File("D:\\javaProgram\\JAVA\\MusicDemo\\web\\"+music.getUrl()+".mp3");
-            if(file.delete()){
-                return_map.put("msg",true);
+                File file = new File("D:\\javaProgram\\JAVA\\MusicDemo\\web\\"+music.getUrl()+".mp3");
+                if(file.delete()){
+                    return_map.put("msg",true);
+                }else {
+                    return_map.put("msg",false);
+
+                }
+
             }else {
                 return_map.put("msg",false);
 
             }
-
         }else {
             return_map.put("msg",false);
 
         }
+
+
+
+
 
         //将map返回给前端
         ObjectMapper mapper = new ObjectMapper();
